requirements.txt
fastapi
uvicorn[standard]
httpx
anthropic
slack-sdk
certifi
Pillow
pydantic-settings
python-dotenv

